Update Download index with regexps
In order to support creating the download index for both liblouis and
utdml we need to get away from an immutable, static, template based
approach.

Instead we now change the version numbers in the right section in the
actual file. This is ugly as hell (nothing that a little regexp
couldn't do) but it does the trick and support both creating releases
for liblouis and utdml.