Inject the toc only after the markdown has been generated
The toc is generated by github pages (when converting the markdown to
html on the server). We trigger this with some magic in the markdown,
namely

* Will be replaced with the ToC
{:toc}

In the olden days we inserted this in the org mode content and used a
hack to make sure this survived the conversion to markdown by pandoc.

But this hack doesn't seem to work anymore. So now we add it after the
conversion, i.e. directly into the generated markdown.